hf_space/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
numpy<2
torch==2.1.0+cpu
torchvision==0.16.0+cpu
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
Pillow>=10.0.0
pyyaml>=6.0
